Search page

+++ app.py
@@ -162,18 +162,21 @@ def edit(id):
         return redirect(url_for('index'))
     return render_template('edit.html', item=item)
 
-@app.route('/delete/<id>', methods=['POST'])
+@app.route('/delete/<id>', methods=['GET', 'POST'])
 @login_required
 def delete(id):
-    db.projects.delete_one({"_id": ObjectId(id)})
-    return redirect(url_for('index'))
+    item = db.projects.find_one({"_id": ObjectId(id)})
+    if request.method == 'POST':
+        db.projects.delete_one({"_id": ObjectId(id)})
+        return redirect(url_for('index'))
+    return render_template('delete.html', item=item)
 
 @app.route('/search')
 @login_required
 def search():
     query = request.args.get('q')
     if query:
-        results = db.projects.find({"name": {"$regex": query, "$options": "i"}})
+        results = db.projects.find({"repo_name": {"$regex": query, "$options": "i"}})
     else:
         results = []
     return render_template('search.html', results=results)
